android [nfc]: Move generic app-status code out of notifications/
